Moved refresh stream button to the bottom left, with a new icon.
"resume" which is now renamed to "restart" was so big that pushing it was inevitable and annoying. I've moved it to the lower left for those who want to use it to get back in sync with the stream. Made this change for myself mostly, but I think many would welcome the change.